Create Using Set
Write a program in the programming language of your choice, which receives the number of students from the user and for that number receives their names and stores them in a set.

Then, print out the list of the names for the students.

At last, ask the user how many students he/she wants to delete from the set, and for that number ask the names from the user and delete those names from the list. Print out the set at the end.