Create Using a matrix to save and store a graph
Write a program in the programming language of your choice to receive the number of nodes in the graph from the user. Then create a matrix for that graph and ask the user to enter the links between nodes. The distance between these nodes also needs to receive from the user and stored in the matrix.